requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.0
Django==4.2.11
django-channels==0.7.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==12.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
mixer==7.2.2
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
priority==2.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
wsproto==1.2.0
